requirements.txt
# Web framework
python-fasthtml==0.12.30
monsterui==1.0.30

# Auth
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Server
uvicorn==0.38.0
watchfiles==1.1.1
python-multipart==0.0.20

# Scraper
playwright==1.55.0
requests==2.32.5
beautifulsoup4==4.14.2

# Utilities
python-dotenv==1.1.1
